DeepJanus-MNIST/requirements.txt
cairocffi==0.9.0
CairoSVG==2.2.1
cffi==1.11.5
deap==1.2.2
grpcio==1.16.1
h5py==2.9.0
Keras-Preprocessing==1.1.0
matplotlib==3.2.1
pycparser==2.19
Pygments==2.3.1
scikit-learn==0.21.2
scipy==1.4.1
tensorflow==2.1.0
